@@ -159,13 +159,13 @@ public class TimePicker extends AnchorPane {
 			public void changed(final ObservableValue<? extends String> property, final String oldValue, final String newValue) {
 				// Wert auf gültigen Bereich beschränken
 				try {
-					final long value = Long.parseLong(newValue);
-					if (value < MIN_VALUE) {
+					final long parsedNumber = Long.parseLong(newValue);
+					if (parsedNumber < MIN_VALUE) {
 						textProperty.setValue(getTwoDigitValue(MIN_VALUE));
-					} else if (value > maxValue) {
+					} else if (parsedNumber > maxValue) {
 						textProperty.setValue(getTwoDigitValue(maxValue));
 					} else if (newValue.length() > 2) {
-						textProperty.setValue(getTwoDigitValue(value));
+						textProperty.setValue(getTwoDigitValue(parsedNumber));
 					} else {
 						textProperty.setValue(newValue);
 					}
